fix utility.h

+++ include/utils/utility.h
@@ -27,7 +27,7 @@ int constraint(int value, int left, int right);
 template <typename Container>
 int countDelta(BipartiteGraph& graph, std::pair<int, int>& range, Container oldOrder, Container newOrder);
 
-int shiftPartialOrder(BipartiteGraph& graph, std::vector<int>::iterator begin, std::vector<int>::iterator end, bool right);
+long long shiftPartialOrder(BipartiteGraph& graph, std::vector<int>::iterator begin, std::vector<int>::iterator end, bool right);
 
 std::vector<int> applyMedianHeuristic(BipartiteGraph *graph);
 std::vector<int> applyBarycentricHeuristic(BipartiteGraph *graph);
